Dedupe plan-review zero-findings test fake to fix R0801 CI warning
The #5032 zero-findings degraded-panel _run_cli fake was copy-pasted
identically into test_plan_review.py and test_plan_review_round.py,
tripping the dedicated pylint duplicate-code (R0801) CI job. Extract it
into a shared make_zero_findings_plan_review_fake_cli helper in
test_support.py and call it from both sites.
