+++ python/test_support.py
@@ -5,10 +5,11 @@ from __future__ import annotations
 import os
 import subprocess
 import sys
-from collections.abc import Mapping, Sequence
+from collections.abc import Callable, Mapping, Sequence
 from dataclasses import dataclass, field
 from pathlib import Path
 
+import collect_results
 from proc import CommandResult
 from run_context import RunContext
 
@@ -122,3 +123,48 @@ def merge_admin_responses(*, double_open_view: bool = False) -> list[CommandResu
     """
     opens = [gh_pr_view(PR_VIEW_OPEN_JSON) for _ in range(2 if double_open_view else 1)]
     return [*opens, gh_pr_view(PR_VIEW_BEHIND_JSON), gh_result(("gh", "pr", "merge"))]
+
+
+def make_zero_findings_plan_review_fake_cli(
+    design: Path, reviewer_file: Path
+) -> Callable[..., subprocess.CompletedProcess[str]]:
+    """Build the shared ``_run_cli`` fake for the #5032 zero-findings degraded-panel path.
+
+    ``test_plan_review`` and ``test_plan_review_round`` both stub the same
+    panel-dispatch / collect-results / aggregate / voter-dispatch / tally sequence
+    for a single OK Cursor reviewer that parses to zero findings. Extracting the
+    identical block here keeps it from tripping the R0801 duplicate-code gate.
+    """
+
+    def fake_run_cli(argv: list[str], *, env: dict[str, str] | None = None) -> subprocess.CompletedProcess[str]:
+        del env
+        if argv[:2] == ["plan-review", "panel-dispatch"]:
+            paths_file = design / "plan-review-panel-paths.txt"
+            _ = (design / "plan-review-slots.ndjson").write_text(
+                '{"slot":"cursor-plan-arch","tool":"cursor","output":"'
+                + str(reviewer_file)
+                + '","prompt_file":"'
+                + str(design / "cursor-plan-arch.prompt")
+                + '"}\n',
+                encoding="utf-8",
+            )
+            _ = paths_file.write_text(str(reviewer_file) + "\n", encoding="utf-8")
+            return subprocess.CompletedProcess(argv, 0, f"PANEL_PRUNED_EMPTY=false\nPANEL_PATHS_FILE={paths_file}\n", "")
+        if argv[:2] == ["agent", "collect-results"]:
+            record = collect_results.CollectorRecord(
+                reviewer_file=str(reviewer_file),
+                tool="cursor",
+                status="OK",
+                exit_code="0",
+            )
+            blocks = ["\n".join(record.fields())]
+            return subprocess.CompletedProcess(argv, 0, "\n\n".join(blocks) + "\n", "")
+        if argv[:2] == ["review", "aggregate-findings"]:
+            return subprocess.CompletedProcess(argv, 0, "REASON=insufficient-input\nAGGREGATED=false\n", "")
+        if argv[:2] == ["plan-review", "voter-dispatch"]:
+            return subprocess.CompletedProcess(argv, 0, "DISPATCH_OK=false\nDEGRADED_PANEL=1\n", "")
+        if argv[:2] == ["plan-review", "tally"]:
+            return subprocess.CompletedProcess(argv, 0, "TALLY_PLAN_REVIEW_STATUS=ok\n", "")
+        return subprocess.CompletedProcess(argv, 0, "", "")
+
+    return fake_run_cli
